Move handler to platform-agnostic `mdm.go` file instead of `microsoft_mdm.go`
