+++ server/service/microsoft_mdm.go
@@ -1635,156 +1635,3 @@ func NewSyncMLCmdStatus(msgRef string, cmdRef string, cmdOrig string, statusCode
 		Items:   nil,
 	}
 }
-
-////////////////////////////////////////////////////////////////////////////////
-// Enqueue Microsoft MDM Command
-////////////////////////////////////////////////////////////////////////////////
-
-type runMDMCommandRequest struct {
-	Command   string   `json:"command"`
-	DeviceIDs []string `json:"device_ids"`
-}
-
-type runMDMCommandResponse struct {
-	*fleet.CommandEnqueueResult
-	Platform string `json:"platform"`
-	Err      error  `json:"error,omitempty"`
-}
-
-func (r runMDMCommandResponse) error() error { return r.Err }
-
-func runMDMCommandEndpoint(ctx context.Context, request interface{}, svc fleet.Service) (errorer, error) {
-	req := request.(*runMDMCommandRequest)
-	result, err := svc.RunMDMCommand(ctx, req.Command, req.DeviceIDs)
-	if err != nil {
-		return runMDMCommandResponse{Err: err}, nil
-	}
-	return runMDMCommandResponse{
-		CommandEnqueueResult: result,
-	}, nil
-}
-
-var (
-	appleMDMPremiumCommands = map[string]bool{
-		"EraseDevice": true,
-		"DeviceLock":  true,
-	}
-
-	microsoftMDMPremiumCommands = map[string]bool{
-		"Wipe": true,
-	}
-)
-
-func (svc *Service) RunMDMCommand(ctx context.Context, rawBase64Cmd string, deviceIDs []string) (result *fleet.CommandEnqueueResult, err error) {
-	// TODO(mna): epic mentions "GitOps users can run commands", check that it is what we want
-	// (how can gitops run a command?). This complicates a lot of things, as gitops has no read
-	// permissions at all (e.g. HostByIdentifier call in fleetctl, list permission here, etc.).
-
-	hosts, err := svc.authorizeAllHostsTeams(ctx, deviceIDs, false, fleet.ActionWrite, &fleet.MDMMicrosoftCommandAuthz{})
-	if err != nil {
-		return nil, err
-	}
-	if len(hosts) == 0 {
-		return nil, newNotFoundError()
-	}
-
-	platforms := make(map[string]bool)
-	for _, h := range hosts {
-		// TODO(mna): hosts lite does not have mdm info
-		if !h.MDMInfo.IsFleetEnrolled() {
-			err := fleet.NewInvalidArgumentError("device_ids", "Can't run the MDM command because one or more hosts have MDM turned off. Run the following command to see a list of hosts with MDM on: fleetctl get hosts --mdm.").WithStatus(http.StatusPreconditionFailed)
-			return nil, ctxerr.Wrap(ctx, err, "check host mdm enrollment")
-		}
-		platforms[h.FleetPlatform()] = true
-	}
-	if len(platforms) != 1 {
-		err := fleet.NewInvalidArgumentError("device_ids", "All hosts must be on the same platform.")
-		return nil, ctxerr.Wrap(ctx, err, "check host platform")
-	}
-
-	// it's a for loop but at this point it's guaranteed that the map has a single value.
-	var commandPlatform string
-	for platform := range platforms {
-		commandPlatform = platform
-	}
-	if commandPlatform != "windows" && commandPlatform != "darwin" {
-		err := fleet.NewInvalidArgumentError("device_ids", "Invalid platform. You can only run MDM commands on Windows or macOS hosts.")
-		return nil, ctxerr.Wrap(ctx, err, "check host platform")
-	}
-
-	// We're supporting both padded and unpadded base64.
-	rawXMLCmd, err := server.Base64DecodePaddingAgnostic(rawBase64Cmd)
-	if err != nil {
-		err = fleet.NewInvalidArgumentError("command", "unable to decode base64 command").WithStatus(http.StatusBadRequest)
-		return nil, ctxerr.Wrap(ctx, err, "decode base64 command")
-	}
-	_ = rawXMLCmd
-
-	/*
-		// a command is a SyncML message
-		var cmdMsg fleet.SyncML
-		if err := xml.Unmarshal(rawXMLCmd, &cmdMsg); err != nil {
-			err = fleet.NewInvalidArgumentError("command", fmt.Sprintf("The payload isn't valid XML. Please provide a file with valid XML: %v", err))
-			return nil, ctxerr.Wrap(ctx, err, "decode SyncML command")
-		}
-		cmd, err := mdm.DecodeCommand(rawXMLCmd)
-		if err != nil {
-			err = fleet.NewInvalidArgumentError("command", "unable to decode plist command").WithStatus(http.StatusUnsupportedMediaType)
-			return 0, nil, ctxerr.Wrap(ctx, err, "decode plist command")
-		}
-	*/
-
-	/*
-		if microsoftMDMPremiumCommands[strings.TrimSpace(cmd.Command.RequestType)] {
-			lic, err := svc.License(ctx)
-			if err != nil {
-				return nil, ctxerr.Wrap(ctx, err, "get license")
-			}
-			if !lic.IsPremium() {
-				return nil, fleet.ErrMissingLicense
-			}
-		}
-	*/
-
-	/*
-			if err := svc.mdmAppleCommander.EnqueueCommand(ctx, deviceIDs, string(rawXMLCmd)); err != nil {
-				// if at least one UUID enqueued properly, return success, otherwise return
-				// error
-				var apnsErr *apple_mdm.APNSDeliveryError
-				var mysqlErr *mysql.MySQLError
-				if errors.As(err, &apnsErr) {
-					if len(apnsErr.FailedUUIDs) < len(deviceIDs) {
-						// some hosts properly received the command, so return success, with the list
-						// of failed uuids.
-						return http.StatusOK, &fleet.CommandEnqueueResult{
-							CommandUUID: cmd.CommandUUID,
-							RequestType: cmd.Command.RequestType,
-							FailedUUIDs: apnsErr.FailedUUIDs,
-						}, nil
-					}
-					// push failed for all hosts
-					err := fleet.NewBadGatewayError("Apple push notificiation service", err)
-					return http.StatusBadGateway, nil, ctxerr.Wrap(ctx, err, "enqueue command")
-
-				} else if errors.As(err, &mysqlErr) {
-					// enqueue may fail with a foreign key constraint error 1452 when one of
-					// the hosts provided is not enrolled in nano_enrollments. Detect when
-					// that's the case and add information to the error.
-					if mysqlErr.Number == mysqlerr.ER_NO_REFERENCED_ROW_2 {
-						err := fleet.NewInvalidArgumentError(
-							"device_ids",
-							fmt.Sprintf("at least one of the hosts is not enrolled in MDM: %v", err),
-						).WithStatus(http.StatusConflict)
-						return http.StatusConflict, nil, ctxerr.Wrap(ctx, err, "enqueue command")
-					}
-				}
-
-				return http.StatusInternalServerError, nil, ctxerr.Wrap(ctx, err, "enqueue command")
-			}
-		return &fleet.CommandEnqueueResult{
-			CommandUUID: cmd.CommandUUID,
-			RequestType: cmd.Command.RequestType,
-		}, nil
-	*/
-	return &fleet.CommandEnqueueResult{}, nil
-}
